fix(social-platforms): #427 — edit form drops provider-specific api_config values (Shiprocket email)
The admin "edit external platform" form couldn't save the Shiprocket account
email (or password / pickup_location): its `mergeApiConfig` shipping branch only
knew Delhivery's fields, its schema omitted `email`/`pickup_location`, and its
defaults never pre-filled them. So edits to a shipping platform's email silently
no-op'd — which surfaced as a 500 on the pickup-registration API (Shiprocket
auth 403 from the stale/wrong email).

Root cause was drift between the create form's `buildApiConfig` and the edit
form's hand-maintained `mergeApiConfig`. Fix it generally so it can't recur and
scales to any future provider:

- New `social-platforms/api-config.ts`: single source of truth — `buildApiConfig`,
  `inferAuthType`, `parseCountryCodes`, and a generic `getFormDefaultsFromApiConfig`
  that maps any existing (redacted) api_config back into flat form defaults.
- create + edit forms both consume it (local duplicates deleted).
- Edit schema gains `email`/`pickup_location` + `.passthrough()` so a new
  provider's fields flow through without a schema edit; `buildApiConfig` still
  scopes which keys persist per category.
- Edit submit now overlays the built config onto the EXISTING one
  (`{...existing, ...overlay}`) so fields the form doesn't render (WhatsApp
  templates/metadata, etc.) survive the wholesale api_config replace; blank
  secrets are restored server-side by `preserveExistingSecrets`.
- Removed a stray debug `console.log` in the edit submit.

Verified: round-trip create→GET(redacted)→PATCH(email changed, password
omitted)→GET persists the new email and preserves the password. tsc unchanged
(70 baseline, 0 new).
