+++ apps/backend/src/admin/components/social-platforms/create-social-platform-component.tsx
@@ -9,6 +9,7 @@ import { KeyboundForm } from "../utilitites/key-bound-form"
 import { Form } from "../common/form"
 import { useRouteModal } from "../modal/use-route-modal"
 import { CategoryProviderFields, hasProviderFields } from "./category-provider-fields"
+import { buildApiConfig, inferAuthType } from "./api-config"
 
 const CreateSocialPlatformSchema = z.object({
   name: z.string().min(1, "Name is required"),
@@ -101,194 +102,6 @@ const CreateSocialPlatformSchema = z.object({
 })
 
 type CreateSocialPlatformForm = z.infer<typeof CreateSocialPlatformSchema>
-
-/**
- * Parse comma-separated country codes into a normalized array of E.164
- * prefixes (e.g. "91, +61" → ["+91", "+61"]). Returns undefined for empty
- * input so api_config doesn't carry an empty array.
- */
-function parseCountryCodes(input: string | undefined | null): string[] | undefined {
-  if (!input) return undefined
-  const codes = input
-    .split(",")
-    .map((s) => s.trim())
-    .filter(Boolean)
-    .map((s) => (s.startsWith("+") ? s : `+${s}`))
-  return codes.length ? codes : undefined
-}
-
-/** Build api_config from form data based on category */
-function buildApiConfig(category: string, data: Record<string, any>): Record<string, any> {
-  const config: Record<string, any> = { provider: data.provider_type }
-
-  switch (category) {
-    case "email":
-      if (data.provider_type === "imap") {
-        Object.assign(config, {
-          host: data.host,
-          port: data.port || 993,
-          user: data.username,
-          password: data.password,
-          tls: data.tls !== false,
-          mailbox: data.mailbox || "INBOX",
-        })
-      } else {
-        Object.assign(config, {
-          api_key: data.api_key,
-          webhook_signing_secret: data.webhook_signing_secret,
-          inbound_domain: data.inbound_domain,
-        })
-      }
-      break
-
-    case "communication":
-      Object.assign(config, {
-        phone_number_id: data.phone_number_id,
-        waba_id: data.waba_id,
-        access_token: data.access_token,
-        app_secret: data.app_secret,
-        webhook_verify_token: data.webhook_verify_token,
-        label: data.label,
-        country_codes: parseCountryCodes(data.country_codes),
-        is_default: data.is_default === true ? true : undefined,
-      })
-      break
-
-    case "sms":
-      if (data.provider_type === "twilio") {
-        Object.assign(config, {
-          account_sid: data.account_sid,
-          auth_token: data.auth_token,
-          from_number: data.from_number,
-          messaging_service_sid: data.messaging_service_sid,
-        })
-      } else {
-        Object.assign(config, {
-          api_key: data.api_key,
-          originator: data.originator,
-        })
-      }
-      break
-
-    case "payment":
-      Object.assign(config, {
-        mode: data.mode || "test",
-        api_key: data.api_key,
-        secret_key: data.secret_key,
-        publishable_key: data.publishable_key,
-        webhook_secret: data.webhook_secret,
-        client_id: data.client_id,
-        client_secret: data.client_secret,
-      })
-      break
-
-    case "shipping":
-      if (data.provider_type === "shiprocket") {
-        // Shiprocket authenticates with email/password (JWT) and references a
-        // registered pickup-location name; no api_key/account_number.
-        Object.assign(config, {
-          mode: data.mode || "test",
-          email: data.email,
-          password: data.password,
-          pickup_location: data.pickup_location,
-        })
-      } else {
-        Object.assign(config, {
-          mode: data.mode || "test",
-          api_key: data.api_key,
-          api_secret: data.api_secret,
-          account_number: data.account_number,
-        })
-      }
-      break
-
-    case "analytics":
-      Object.assign(config, {
-        tracking_id: data.tracking_id,
-        api_key: data.api_key,
-        api_secret: data.api_secret,
-        project_token: data.project_token,
-        host: data.host,
-      })
-      break
-
-    case "storage":
-      Object.assign(config, {
-        access_key_id: data.access_key_id,
-        secret_access_key: data.secret_access_key,
-        bucket: data.bucket,
-        region: data.region,
-        endpoint: data.endpoint,
-        cloud_name: data.cloud_name,
-        api_key: data.api_key,
-        api_secret: data.api_secret,
-        project_id: data.project_id,
-      })
-      break
-
-    case "crm":
-      Object.assign(config, {
-        api_key: data.api_key,
-        client_id: data.client_id,
-        client_secret: data.client_secret,
-        instance_url: data.instance_url,
-        portal_id: data.portal_id,
-      })
-      break
-
-    case "authentication":
-      Object.assign(config, {
-        domain: data.domain,
-        client_id: data.client_id,
-        client_secret: data.client_secret,
-        audience: data.audience,
-        secret_key: data.secret_key,
-        publishable_key: data.publishable_key,
-        project_id: data.project_id,
-        api_key: data.api_key,
-      })
-      break
-
-    case "google":
-      // Credentials (client_id/client_secret/developer_token) are managed in
-      // the per-row Google Business Panel after creation. The create form
-      // just gets an empty api_config — strip the default `provider` key.
-      return {}
-  }
-
-  // Remove undefined/empty values
-  return Object.fromEntries(
-    Object.entries(config).filter(([_, v]) => v !== undefined && v !== "")
-  )
-}
-
-/** Infer auth_type from category and provider */
-function inferAuthType(category: string, providerType?: string): string {
-  switch (category) {
-    case "email":
-      return providerType === "resend" ? "api_key" : "basic"
-    case "communication":
-      return "bearer"
-    case "sms":
-      return providerType === "twilio" ? "basic" : "api_key"
-    case "payment":
-      return "api_key"
-    case "shipping":
-      return providerType === "shiprocket" ? "basic" : "api_key"
-    case "analytics":
-      return "api_key"
-    case "storage":
-      return "api_key"
-    case "crm":
-      return providerType === "hubspot" ? "api_key" : "oauth2"
-    case "authentication":
-      return "oauth2"
-    case "google":
-      return "oauth2"
-    default:
-      return "api_key"
-  }
-}
 
 export const CreateSocialPlatformComponent = () => {
   const { handleSuccess } = useRouteModal();
